refactor: home components folder
