+++ components/home/SubjectClassesDialog.tsx
@@ -2,7 +2,7 @@
 import ConfirmDeleteDialog from "@/components/common/ConfirmDeleteDialog";
 import MultilangText from "@/components/common/MultilingualText";
 import BrandIcon from "@/components/icons/BrandIcon";
-import ClassroomSubjectDialog from "@/components/subject/ClassroomSubjectDialog";
+import ClassroomSubjectDialog from "@/components/home/ClassroomSubjectDialog";
 import SnackbarContext from "@/contexts/SnackbarContext";
 import deleteClassroomSubject from "@/utils/backend/subject/deleteClassroomSubject";
 import getClassroomSubjectsOfSubject from "@/utils/backend/subject/getClassroomSubjectsOfSubject";
